fw: add ARMv8 MPU partial support
Adjust MPU driver so that it is kind of functional on ARMv8 MPU.
However, memory layout code relies on implementation details of the
ARMv7 MPU which are not available, so the code actually needs lots of
love to find a better solution to this mess. Also, SF32LB has some
extra MPU requirements that need more adjustments.

Summarizing, MPU is enabled but in SiFli HAL glue code
(system_bf0_ap.c), but obviously not respecting any of the original
PebbleOS memory layouts (e.g. all RAM is RWX by anyone).

Signed-off-by: Gerard Marull-Paretas <gerard@teslabs.com>