drivers/task_watchdog: use board-specified RTC for watchdog on nRF5, and use 60Hz RTC on Asterix
We're out of RTCs on Asterix, so we're going to have to repurpose the
watchdog RTC to trigger the GPIOTE for toggling the display VCOM.  Such
is our lot in life, I suppose.  Make the nRF5 watchdog explicitly
configured in the board.h (so you can't forget about "something on the
system" using a system resource that isn't specified there...), and
prepare to run it at 60Hz for VCOM usage.

Signed-off-by: Joshua Wise <joshua@accelerated.tech>